chore(stack): bloco C -- wgpu 29, vello 0.10, parley 0.11, e um vies MORTO
As seis crates de GPU e texto sobem JUNTAS, como o bloco exigia:
wgpu 28 -> 29.0.4 · vello 0.8 -> 0.10.0 · naga 28 -> 29.0.4 ·
skrifa 0.40 -> 0.44.0 · parley 0.6 -> 0.11.1 · fontique 0.6 -> 0.11.1 ·
accesskit fixado em 0.24.1. Mais o usvg 0.43 -> 0.48 (a F16 entrou aqui porque
e' ela que solta o kurbo).

## ⭐⭐⭐ O achado: 48 gates cairam por um valor INERTE

A fotografia dos gates de GPU foi de 12 vermelhos para 65, e quase toda a suite
do mesh-render estava la'. Causa, uma linha:

  In Device::create_render_pipeline, label = 'ph2d-mesh wire'
    Depth bias is not compatible with non-triangle topology LineList

O passe de arestas declarava `DepthBiasState { constant: -4, slope_scale: -1.0 }`
com um comentario longo a explicar que "o vies negativo e' o que faz a aresta
ganhar da face". O raciocinio esta' certo PARA TRIANGULOS; a topologia e'
`LineList`. O WebGPU exige vies zero fora de topologia de triangulos e o Vulkan
aplica-o so' a poligonos ⇒ aquele valor NUNCA foi aplicado por backend nenhum.

⭐⭐ E esta casa ja' o tinha medido. A sonda `probe_wire_continuity` carrega,
escrito: "nao ha' um gate afirmando «o vies do pipeline nao alcanca uma linha»,
que e' o achado mais caro desta investigacao -- porque eu nao consigo faze^-lo
falhar pelo motivo que ele alegaria", com a varredura ao lado: `constant` de 0 a
-4096, tinta IDENTICA ao pixel. A cura verdadeira foi construida no shader
(`WIRE_DEPTH_NUDGE` + o empurrao lateral). O campo do pipeline ficou para tras
porque NADA o obrigava a sair.

⇒ o wgpu 29 obrigou: ele valida o que o 28 tolerava calado.
⚠️ Um valor inerte nao custa nada ate' ao dia em que a plataforma deixa de o
tolerar; nesse dia custa a suite inteira -- e enquanto la' esta', quem o le^ pela
primeira vez acredita no comentario.
⚠️ O gate que faltava agora existe e NAO e' nosso: e' a validacao do wgpu. Quem
repuser um vies ali nao ve^ um teste vermelho -- ve^ a malha 3D deixar de desenhar.

## ⛔ Quatro tarefas do plano mandavam trabalho INEXISTENTE

  C10  34 edicoes de `VertexState::buffers`  -> 0 (a linha e' byte a byte igual)
  C13  3 varyings a precisar de `flat`       -> 0 (a regra ja' existia no naga 28,
                                                   e os 3 sao atributos de vertice)
  C16  renomes de `Alignment`                -> 0 (feitos antes da 0.6)
  C17  `peniko::Font` -> `FontData`          -> 0 (idem)
  C22  "o gradiente mudou de espaco"         -> FALSO, e' byte-identico

## ⚠️ E a contagem errou nas duas receitas que sobraram

  bind_group_layouts no ph2d-render      plano 9 -> medido 13
  idem nas crates-ferramenta             plano 8/9 -> medido 10

Os agentes nao confiaram no numero: varreram e reportaram. Uma varredura da
arvore toda (bind_group_layouts sem Some, depth_* sem Some, SurfaceError,
.suboptimal) devolve ZERO.

## Cinco quebras de texto que o plano nao previu (todas erro de compilacao)

parley deixou de re-exportar o `swash` · `FontStack` APAGADO (-> `FontFamily`) ·
`FontSettings<T>` partido em dois tipos · `FontVariation.tag` de u32 para `Tag` ·
`Layout::align` perdeu um argumento (era a MESMA largura escrita duas vezes).
E uma que NAO e' erro: `Limits::max_*_buffer_binding_size` u32 -> u64.

## ⭐ Duas curas melhores que o original

  · `AcquireError::Occluded` ja' existia e o shell ja' o tratava -- o wgpu 28 nao
    tinha a situacao, logo nenhum produtor podia emiti-lo. O 29 tem. Restaura.
  · `fontique::Script` mudou de dono e perdeu o `From<&str>`, que era
    `try_into().unwrap_or_default()`: um codigo de escrita com tamanho errado
    virava quatro zeros EM SILENCIO. O substituto (`from_bytes([u8;4])`) torna
    isso erro de compilacao.
  ⚠️ Esta segunda o reconhecimento declarou INEXISTENTE -- ele conferiu
  assinaturas de FUNCAO e nao a origem do TIPO.

## ⚠️ O que muda PIXEL (compila verde -- e' a lista do smoke do dono)

  1. toda imagem do vello desloca-se MEIO PIXEL (correccao de defeito upstream:
     passou a amostrar no centro). Mais visivel no modo de arte de pixel: pode
     aparecer deslocada uma coluna inteira.
  2. `ImageQuality::High` era bilinear disfarcado e passa a bicubico de verdade.
  3. o `y_offset` do glifo inverteu -- e isso CURA um defeito nosso: acentos por
     mark-attachment desciam em vez de subir. Latino precomposto nao muda um bit.
  4. o shaper saltou harfrust 0.3 -> 0.12: texto pode mover-se um pixel.

## Colapso de duplicacao (de graca)

  skrifa 3 copias -> 1 · read-fonts 3 -> 1 · swash 1 -> 0 (saiu da arvore) ·
  kurbo 2 -> 1 (o usvg) · wgpu-types 2 -> 1 (esse foi o bloco D)

## Portao

  cargo check --workspace --all-targets   verde
  clippy --workspace --all-targets        exit 0, zero avisos
  fmt                                     limpo
  suite                                   20 044 / 20 044
  gates de GPU (que o CI nunca roda)      12 vermelhos ANTES -> 10 DEPOIS
                                          os 4 que mudaram sao sondas de relogio;
                                          o unico "novo" passa 3/3 sozinho a carga
                                          0,36. ⇒ ZERO regressoes de GPU.

⚠️ ABERTO: `parley 0.11.1` tem feature nova `complex-scripts` NAO-default -- sem
ela, quebra de linha degrada para tailandes/khmer/lao/birmanes. Nenhum idioma do
produto hoje; fica NOMEADO para nao passar por omissao.

+++ crates/ph2d-gpu-cook/src/lib.rs
@@ -626,7 +626,8 @@ impl GpuCook {
         // bind group turns it into a validation panic.
         let instance_bytes =
             u64::from(count) * std::mem::size_of::<ph2d_render::RenderInstance>() as u64;
-        let binding_limit = u64::from(gpu.device.limits().max_storage_buffer_binding_size);
+        // wgpu 29: `max_storage_buffer_binding_size` is already `u64`.
+        let binding_limit = gpu.device.limits().max_storage_buffer_binding_size;
         if instance_bytes > binding_limit {
             return Err(GpuCookError::BindingTooLarge {
                 bytes: instance_bytes,
